Rename exceptions.ts in evm to errors.ts to be uniform with eof errors file
