+++ packages/evm/src/index.ts
@@ -1,6 +1,6 @@
 import { EOFContainer, validateEOF } from './eof/container.ts'
+import { EVMError as EVMErrorMessage, EvmError } from './errors.ts'
 import { EVM } from './evm.ts'
-import { EVMError as EVMErrorMessage, EvmError } from './exceptions.ts'
 import { Message } from './message.ts'
 import { getOpcodesForHF } from './opcodes/index.ts'
 import {
