+++ packages/evm/src/precompiles/08-bn254-pairing.ts
@@ -1,7 +1,7 @@
 import { bytesToHex } from '@ethereumjs/util'
 
+import { EVMError, EvmError } from '../errors.ts'
 import { EvmErrorResult, OOGResult } from '../evm.ts'
-import { EVMError, EvmError } from '../exceptions.ts'
 
 import { getPrecompileName } from './index.ts'
 import { gasLimitCheck, moduloLengthCheck } from './util.ts'
